LRQA-13777 Move BaseFunction.java and BaseMacro.java back to original location
